explicit dtype denoted for torch.tensors of input_dict

+++ source/masif_modules/train_masif_site.py
@@ -210,10 +210,10 @@ def train_masif_site(
                     "theta_coords": torch.tensor(theta_wrt_center, dtype=torch.float32),
                     "input_feat": torch.tensor(input_feat, dtype=torch.float32),
                     "mask": torch.tensor(mask, dtype=torch.float32),
-                    "labels": torch.tensor(iface_labels_dc, dtype=torch.long),
-                    "pos_idx": torch.tensor(pos_labels, dtype=torch.long),
-                    "neg_idx": torch.tensor(neg_labels, dtype=torch.long),
-                    "indices_tensor": torch.tensor(indices, dtype=torch.long),
+                    "labels": torch.tensor(iface_labels_dc, dtype=torch.int32),
+                    "pos_idx": torch.tensor(pos_labels, dtype=torch.int32),
+                    "neg_idx": torch.tensor(neg_labels, dtype=torch.int32),
+                    "indices_tensor": torch.tensor(indices, dtype=torch.int32),
                 }
                 # move input tensors to the same device w/ parameter tensors of the model
                 input_dict = {key: tensor.to(device) for key, tensor in input_dict.items()}
@@ -293,10 +293,10 @@ def train_masif_site(
                     "theta_coords": torch.tensor(theta_wrt_center, dtype=torch.float32),
                     "input_feat": torch.tensor(input_feat, dtype=torch.float32),
                     "mask": torch.tensor(mask, dtype=torch.float32),
-                    "labels": torch.tensor(iface_labels_dc, dtype=torch.long),
-                    "pos_idx": torch.tensor(pos_labels, dtype=torch.long),
-                    "neg_idx": torch.tensor(neg_labels, dtype=torch.long),
-                    "indices_tensor": torch.tensor(indices, dtype=torch.long),
+                    "labels": torch.tensor(iface_labels_dc, dtype=torch.int32),
+                    "pos_idx": torch.tensor(pos_labels, dtype=torch.int32),
+                    "neg_idx": torch.tensor(neg_labels, dtype=torch.int32),
+                    "indices_tensor": torch.tensor(indices, dtype=torch.int32),
                 }
                 # move input tensors to the same device w/ parameter tensors of the model
                 input_dict = {key: tensor.to(device) for key, tensor in input_dict.items()}
